Autocomplete: split `createProvider` into multiple files (#5356)
- Moving code around to make `createProvider` less intimidating and
easier to scan:
    - Extracted model parsing logic into a separate module
    - Extracted experiment-model resolution into a separate module.
- No functional changes.
